% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.vertGLM.R
\name{coef.ds.glm}
\alias{coef.ds.glm}
\title{Coefficients Method for ds.glm Objects}
\usage{
\method{coef}{ds.glm}(object, ...)
}
\arguments{
\item{object}{A ds.glm object}

\item{...}{Additional arguments (ignored)}
}
\value{
Named numeric vector of coefficients
}
\description{
Extract coefficients from a ds.glm object.
}
