% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_utils.R
\name{.dsvert_set_chunk_size}
\alias{.dsvert_set_chunk_size}
\title{Set the effective chunk size (internal)}
\usage{
.dsvert_set_chunk_size(size)
}
\arguments{
\item{size}{Integer. Chunk size in characters.}
}
\description{
Set the effective chunk size (internal)
}
\keyword{internal}
