% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.vertGLM.k2.R
\name{.lbfgs_direction}
\alias{.lbfgs_direction}
\title{L-BFGS two-loop recursion (Nocedal & Wright, Algorithm 7.4)}
\usage{
.lbfgs_direction(grad, s_hist, y_hist)
}
\description{
L-BFGS two-loop recursion (Nocedal & Wright, Algorithm 7.4)
}
\keyword{internal}
