% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.vertGLM.k2.R
\name{k2-beaver-lbfgs-client}
\alias{k2-beaver-lbfgs-client}
\title{K=2 Beaver L-BFGS Pipeline}
\description{
Secure K=2 GLM with L-BFGS optimizer.
Gaussian: identity link, constant Fisher, L-BFGS (2 rounds/iter).
Binomial/Poisson: wide spline sigmoid/exp, L-BFGS (6 rounds/iter, no Fisher).
}
