% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.vertCor.R
\name{print.ds.cor}
\alias{print.ds.cor}
\title{Print Method for ds.cor Objects}
\usage{
\method{print}{ds.cor}(x, digits = 3, ...)
}
\arguments{
\item{x}{A \code{ds.cor} object.}

\item{digits}{Number of displayed digits.}

\item{...}{Additional arguments passed to \code{print}.}
}
\description{
Prints a summary of the DP correlation result.
}
