% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.vertGLM.R
\name{print.ds.glm}
\alias{print.ds.glm}
\title{Print Method for ds.glm Objects}
\usage{
\method{print}{ds.glm}(x, ...)
}
\arguments{
\item{x}{A ds.glm object}

\item{...}{Additional arguments (ignored)}
}
\description{
Prints a summary of GLM results.
}
