% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.validateIdFormat.R
\name{print.ds.id.validation}
\alias{print.ds.id.validation}
\title{Print Method for ds.id.validation Objects}
\usage{
\method{print}{ds.id.validation}(x, ...)
}
\arguments{
\item{x}{A ds.id.validation object}

\item{...}{Additional arguments (ignored)}
}
\description{
Prints a summary of ID format validation results.
}
