% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.vertGLM.R
\name{summary.ds.glm}
\alias{summary.ds.glm}
\title{Summary Method for ds.glm Objects}
\usage{
\method{summary}{ds.glm}(object, ...)
}
\arguments{
\item{object}{A ds.glm object}

\item{...}{Additional arguments (ignored)}
}
\description{
Prints detailed summary including deviance and fit statistics.
}
