% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.identity.R
\name{ds.getIdentityPks}
\alias{ds.getIdentityPks}
\title{Query Server Identity Public Keys}
\usage{
ds.getIdentityPks(datasources = NULL)
}
\arguments{
\item{datasources}{DataSHIELD connections. If NULL, uses all available.}
}
\value{
Named list: server_name -> identity_pk (base64url string).
}
\description{
Queries all connected DataSHIELD servers for their Ed25519
identity public keys. Used by administrators to discover PKs for
configuring the \code{dsvert.trusted_peers} option on each server.
}
\examples{
\dontrun{
pks <- ds.getIdentityPks(datasources = conns)
# Set trusted peers on each server via opalr:
# dsadmin.set_option(opal, "dsvert.trusted_peers", paste(pks, collapse=","))
}

}
