% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.vertCor.R
\name{ds.vertCor}
\alias{ds.vertCor}
\title{Client function for vertCor for finding correlation of vertically partitioned data in DataSHIELD}
\usage{
ds.vertCor(variable_names, datasources = NULL)
}
\arguments{
\item{variable_names}{names of variables to be used in correlation matrix}

\item{datasources}{conns for the defined study}
}
\value{
correlation matrix
}
\description{
Client function for vertCor for finding correlation of vertically partitioned data in DataSHIELD
}
\examples{
TODO
}
