% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.vertCor.R
\name{print.ds.cor}
\alias{print.ds.cor}
\title{Print Method for ds.cor Objects}
\usage{
\method{print}{ds.cor}(x, digits = 3, ...)
}
\arguments{
\item{x}{A ds.cor object}

\item{digits}{Number of digits for correlation matrix. Default 3.}

\item{...}{Additional arguments (ignored)}
}
\description{
Prints a summary of correlation results.
}
