% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.vertPCA.R
\name{print.ds.pca}
\alias{print.ds.pca}
\title{Print Method for ds.pca Objects}
\usage{
\method{print}{ds.pca}(x, ...)
}
\arguments{
\item{x}{A ds.pca object}

\item{...}{Additional arguments (ignored)}
}
\description{
Prints a summary of PCA results.
}
